Pull request #42: Story/CIGCPDPH2 933 pods 1 1 update install scripts
Merge in NRCSAPPS/pods from story/CIGCPDPH2-933-pods-1-1-update-install-scripts to develop

Squashed commit of the following:

commit 66c7d204a6f368ff620d237e5c17f7c3095c7e1f
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Tue May 2 17:01:59 2023 -0700

    Added exit 0 to signal success to pipeline.

commit eb9f97f164a8d55e280c3bad3cb6db4eb51cae71
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Tue May 2 16:10:44 2023 -0700

    Adjust sym link handling

commit 05f57479a45183c2a57c62f9ca80397f6ce8334c
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Tue May 2 16:09:03 2023 -0700

    Correction to alive file sym link handling

commit 5df51a0c43f79a4c1c3b9f216f04f7b2267134cc
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Tue May 2 07:53:07 2023 -0700

    Adjust sym link creation only if missing

commit 757b4a65141c72e303c0a7187c2b957626279268
Merge: 48c9d0f8 2e13d38b
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Mon May 1 14:55:58 2023 -0700

    Merge and develop and resolve conflicts:

commit 48c9d0f837b80c52d8372250118b3935b3cf4ed6
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Mon May 1 13:58:02 2023 -0700

    Minor change to 2nd server deploy after comparison with 1st server approach

commit bf5dfffb636f7b30d225f41953ecb25909ff6a3a
Author: William Keller <william.keller@spatialfront.com>
Date:   Mon May 1 15:59:46 2023 -0400

    adjusted limits in shmu form

commit cff717270d2d4795c124f4615fb7669202e4abeb
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Mon May 1 11:18:48 2023 -0700

    fix rconditional emoval of last tarball

commit 4f7bce8bf3de8f631a47525f4b1eee01b5fc608d
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Mon May 1 11:14:57 2023 -0700

    Clear out old build folder and previous tarball to avoid bloat

commit bcc79b1109742aa4ee17b2f65d7fc4e68a869056
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Mon May 1 11:12:41 2023 -0700

    Fix to change rangeland to range

commit 65fc7bd5f7803d604bf1a731522ba208192c6e9b
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Mon May 1 09:47:32 2023 -0700

    Removed farm_update and added select2 in core.extension.yml Mike recommended both. Not sure what removing farm_update was for but select2 was needed for the module to be enabled.

commit 56d15c47680de2a5c7f34ca0bc8c92004e8f6334
Author: William Keller <william.keller@spatialfront.com>
Date:   Fri Apr 28 15:17:01 2023 -0400

    fixed location of eauth check

commit 795e1e02a175928430e978821e00410618741a48
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Fri Apr 28 12:00:11 2023 -0700

    Change to 2nd server deploy eliminating maintenance mode and prompting for alive file rename before continuing for symlink creation.

commit b88a5a9657d902079fce4d226c155ff8322153da
Merge: 6161dbf5 948bf766
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Fri Apr 28 11:43:07 2023 -0700

    Merge branch 'story/CIGCPDPH2-933-pods-1-1-update-install-scripts' of https://bitbucket.fpac.usda.gov/scm/nrcsapps/pods into story/CIGCPDPH2-933-pods-1-1-update-install-scripts

commit 6161dbf5f1907ec870837c4d2f7307d0ac5454dd
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Fri Apr 28 11:42:27 2023 -0700

    Fixed syntax goof in update hook for select2 module

commit 948bf766b748f82798b14301e16d49ac11b91c7c
Author: William Keller <william.keller@spatialfront.com>
Date:   Fri Apr 28 14:26:21 2023 -0400

    bugfix for users without eauth ids; update to  post-login redirect

commit 7d9ac657b56fe92815570e8dc0faabec1582338d
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Fri Apr 28 10:51:26 2023 -0700

    Fix syntax error, else vs elif
    Add wait after composer install so it will finish before the tarball is created

commit 16caa2bf0d9a130e3822e8e8541a514e449cd9cb
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Fri Apr 28 09:01:31 2023 -0700

    Fix typo in line that pulls in the custom modules.

commit d96616ceaab833dbbb88ee77cfc02933cc7e455f
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Fri Apr 28 08:55:18 2023 -0700

    Removing v from the tarball creation for less output

commit 7706f085ce7c01b99026ffc8b60c61e7d08e3470
Merge: afce08a9 d81d90b9
Author: forsyth.bacon@usda.gov <forsyth.bacon@usda.gov>
Date:   Fri Apr 28 07:37:03 2023 -0700

    Merge branch 'develop'

... and 30 more commits

+++ scripts/pods_1.1_deploy_init.sh
@@ -80,8 +80,11 @@ if [[ "$SUPPORTED_ENVS" == *"$2"* ]]; then
 		    wait
 
         else
-            echo "recreate alive file symbolic link"
-            ln -s /app/httpd/htdocs/Alive1.html /app/www/html/$ENVNAME/web/Alive1.html 2>&1|tee -a $LOGFILE
+            # Recreate alive file link if needed
+            if [[ ! -L /app/www/html/$ENVNAME/web/Alive1.html ]]; then
+                echo "recreating alive file symbolic link"
+                ln -s /app/httpd/htdocs/Alive1.html /app/www/html/$ENVNAME/web/Alive1.html 2>&1|tee -a $LOGFILE
+            fi
 
             echo "copy settings.php from backup made earlier" 2>&1|tee -a $LOGFILE
             cp -fp /app/upload/pods.$2.settings.php.bak /app/www/html/"$ENVNAME"/web/sites/default/settings.php 2>&1|tee -a $LOGFILE
